Created stories for sprint cards
