+++ client/src/components/exercise/lab0/DesignLabSection/LabIdeaActivity/DesignLabIntroduction.js
@@ -1,63 +1,31 @@
 import React, { useContext, useEffect, useState } from "react";
-import Lab0Context from "../Lab0Context";
-import DragDropGame from "../../../all-components/DragAndDrop/DragDropGame";
-import _ from "lodash";
+import DragDropGame from "../../../../all-components/DragAndDrop/DragDropGame";
+import Lab0Context from "../../Lab0Context";
 import {
-  columns_new_category,
-  previous_Bank,
-  correctAssignmentsNewCategory,
-} from "../../../../constants/lab0/DesignALab/LabCategoryDND";
+  columns,
+  initial_Bank,
+  correctAssignments,
+} from "../../../../../constants/lab0/DesignALab/LabCategoryDND";
 
-const DesignSortNewCategory = () => {
-  const { newCategoryName, newLabTopics, handleNav } = useContext(Lab0Context);
-
+const DesignLabIntroduction = () => {
   const [success, setSuccess] = useState(false);
-  const [cols, setCols] = useState(() => structuredClone(columns_new_category));
-  const [bank, setBank] = useState(() => structuredClone(previous_Bank));
+  const [cols, setCols] = useState(() => structuredClone(columns));
+  const [bank, setBank] = useState(() => structuredClone(initial_Bank));
   const [correct, setCorrect] = useState(() =>
-    structuredClone(correctAssignmentsNewCategory),
+    structuredClone(correctAssignments),
   );
 
   // Handles if user goes to another page, reset objects
   useEffect(() => {
-    setCols(structuredClone(columns_new_category));
-    setBank(structuredClone(previous_Bank));
-    setCorrect(structuredClone(correctAssignmentsNewCategory));
+    setCols(structuredClone(columns));
+    setBank(structuredClone(initial_Bank));
+    setCorrect(structuredClone(correctAssignments));
   }, []);
 
+  const { handleNav } = useContext(Lab0Context);
   const navigateNext = () => {
+    console.log(columns);
     handleNav("DesignNewCategory");
-  };
-
-  const Columns = cols.map((col) =>
-    col.id === "column3" ? { ...col, title: newCategoryName } : col,
-  );
-
-  const newBank = () => {
-    newLabTopics.forEach((topic) => {
-      const newId = bank.length + 1;
-      bank.push({
-        id: "card" + newId,
-        content: topic.value,
-        isCorrect: true,
-      });
-    });
-    return bank;
-  };
-
-  const newCorrectAssignments = () => {
-    return correct.map((col) => {
-      if (col.id === "column3") {
-        return {
-          ...col,
-          cards: [
-            ...col.cards,
-            ...newLabTopics.map((_, i) => "card" + (8 + 1 + i)),
-          ],
-        };
-      }
-      return col;
-    });
   };
 
   return (
@@ -87,12 +55,13 @@ const DesignSortNewCategory = () => {
       <br />
       <div className="tw-p-3 tw-h-[60rem]">
         <DragDropGame
+          gameStyle={"tw-flex tw-flex-col tw-w-full"}
           containerStyle={"tw-flex tw-gap-5 tw-p-5 tw-w-full tw-justify-center"}
           colStyle={
             "tw-px-4 tw-rounded-lg tw-shadow-md tw-flex tw-flex-1 tw-min-h-[20rem] tw-pb-4 tw-flex tw-flex-col tw-w-full"
           }
           bankStyle={
-            "tw-grid tw-grid-cols-3 sm:tw-grid-cols-4 lg:tw-grid-cols-5 tw-gap-4 tw-p-4 tw-border-t-4 " +
+            "tw-grid tw-grid-cols-3 xs:tw-grid-cols-3 md:tw-grid-cols-4 xl:tw-grid-cols-5 tw-gap-4 tw-p-4 tw-border-t-4 " +
             "tw-border-dashed tw-border-black tw-border-r-0 tw-border-l-0 tw-border-b-0 tw-my-5 tw-justify-center " +
             "tw-w-full tw-min-w-full tw-max-w-full tw-flex-shrink-0 tw-h-auto tw-min-h-[10rem] " +
             "after:tw-content-[''] after:tw-block after:tw-w-[12rem] after:tw-h-[6rem] after:tw-opacity-0 after:tw-col-span-1"
@@ -102,25 +71,25 @@ const DesignSortNewCategory = () => {
             "tw-text-black tw-w-full tw-flex tw-items-center tw-justify-center tw-h-auto "
           }
           bankCardStyle={
-            "tw-bg-white tw-border-solid tw-border-labBlue tw-rounded-md tw-p-4 tw-m-2 tw-shadow-sm tw-cursor-grab focus:tw-shadow-xl " +
-            "tw-text-black tw-h-[4rem] tw-flex tw-items-center tw-body-text tw-leading-snug xs:tw-text-sm md:tw-text-[1rem] tw-justify-center tw-min-w-[6rem]"
+            "tw-bg-white tw-border-solid tw-border-labBlue tw-rounded-md tw-p-2 tw-m-2 tw-shadow-sm tw-cursor-grab focus:tw-shadow-xl " +
+            "tw-text-black tw-h-[4rem] tw-flex tw-items-center tw-body-text tw-leading-snug xs:tw-text-sm md:tw-text-[1rem] tw-justify-center tw-min-w-[7rem]"
           }
-          msgStyle={`${!success ? "tw-bg-error" : "tw-bg-success"} tw-p-4 tw-my-1 tw-text-white tw-rounded-md tw-mb-8`}
+          msgStyle={`${!success ? "tw-bg-error" : "tw-bg-success"} tw-p-4 tw-my-1 tw-text-white tw-rounded-md tw-mb-3`}
           colHeaderStyle={
             "tw-flex tw-items-center tw-justify-center tw-h-[6rem] tw-text-center tw-px-4 tw-py-4 " +
             "tw-mx-2 tw-my-2 tw-bg-labYellow tw-rounded-md tw-font-bold"
           }
-          colContainerStyle={"tw-w-1/3"}
-          cols={Columns}
-          initial_bank={_.shuffle(newBank())}
-          correct_assignments={newCorrectAssignments()}
+          cols={cols}
+          initial_bank={bank}
+          correct_assignments={correct}
           setSuccess={setSuccess}
           sucess={success}
           handleNav={navigateNext}
+          colContainerStyle={"tw-w-1/3"}
         />
       </div>
     </div>
   );
 };
 
-export default DesignSortNewCategory;
+export default DesignLabIntroduction;
