+++ client/src/components/exercise/lab0/Main.js
@@ -7,12 +7,12 @@ import ExperientialIntroduction from "./DesignLabSection/ExperientialActivity/Ex
 import CreateExperientialExercise from "./DesignLabSection/ExperientialActivity/CreateExperientialExercise";
 import DesignLabDecision from "./DesignLabSection/ExperientialActivity/DesignLabDecision";
 import ScrumIntroduction from "./DesignLabSection/ScrumActivity/ScrumIntroduction";
-import DesignLabIntroduction from "./DesignLabSection/DesignLabIntroduction";
+import DesignLabIntroduction from "./DesignLabSection/LabIdeaActivity/DesignLabIntroduction";
 import ScrumBoardActivity from "./DesignLabSection/ScrumActivity/ScrumBoardActivity";
 import ScrumVelocityReading from "./DesignLabSection/ScrumActivity/ScrumVelocityReading";
 import ScrumVelocityActivity from "./DesignLabSection/ScrumActivity/ScrumVelocityActivity";
-import DesignNewCategory from "./DesignLabSection/DesignNewCategory";
-import DesignSortNewCategory from "./DesignLabSection/DesignSortNewCategory";
+import DesignNewCategory from "./DesignLabSection/LabIdeaActivity/DesignNewCategory";
+import DesignSortNewCategory from "./DesignLabSection/LabIdeaActivity/DesignSortNewCategory";
 
 const Main = () => {
   const [exerciseState, setExerciseState] = useState(
