+++ client/src/constants/lab0/DesignALab/ScrumDND.js
@@ -8,19 +8,35 @@ const initialBank = [
   {
     id: "card1",
     content: "Card 1",
-    body: "Velocity: 6",
+    body: `
+1. Create Code block for Improving Focus Order
+2. Create Data Service for Exercise
+3. Create Newly Accessible Webpage
+4. Create Key Takeaways
+    `,
     color: "tw-bg-[#F83F3F]",
   },
   {
     id: "card2",
     content: "Card 2",
-    body: "Velocity: 7",
+    body: `
+1. Create Skeleton for Exercise Structure
+2. Create Exercise Introduction
+3. Display Blurred Webpage
+4. Create Screen Reader
+5. Create Tab through Nav
+6. Create Out of Order Tab Nav
+    `,
     color: "tw-bg-[#6D67E6]",
   },
   {
     id: "card3",
     content: "Card 3",
-    body: "Velocity: 3",
+    body: `
+1. Explanation of Disordered Focus
+2. Hearing Mismatched Labels
+3. Explanation of the Inaccessibility
+    `,
     color: "tw-bg-[#FF860D]",
   },
 ];
